requirements/metric3d_req.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
addict=2.4.0=pyhd8ed1ab_2
alsa-lib=1.2.11=hd590300_1
aom=3.9.0=hac33072_0
attr=2.5.1=h166bdaf_1
blas=1.0=openblas
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.28.1=hd590300_0
ca-certificates=2024.3.11=h06a4308_0
cairo=1.18.0=h3faef2a_0
certifi=2024.2.2=py312h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.2.1=pypi_0
cuda-version=11.8=hcce14f8_3
cudatoolkit=11.8.0=h6a678d5_0
cudnn=8.9.7.29=hbc23b4c_3
cycler=0.12.1=pypi_0
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
double-conversion=3.3.0=h59595ed_0
expat=2.6.2=h6a678d5_0
ffmpeg=6.1.1=gpl_he44c6f3_112
filelock=3.13.1=py312h06a4308_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.2=h14ed4e7_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.53.0=pypi_0
freeglut=3.4.0=h6a678d5_0
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h7b6447c_0
fsspec=2024.3.1=py312h06a4308_0
gettext=0.22.5=h59595ed_2
gettext-tools=0.22.5=h59595ed_2
glib=2.80.2=hf974151_0
glib-tools=2.80.2=hb6ce0ca_0
gmp=6.3.0=h59595ed_1
gnutls=3.7.9=hb077bed_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.24.4=h9ad1361_0
gstreamer=1.24.4=haf2f30d_0
harfbuzz=8.5.0=hfac3d4d_0
hdf5=1.14.3=nompi_hdf9ad27_103
huggingface-hub=0.23.2=pypi_0
icu=73.2=h59595ed_0
idna=3.7=py312h06a4308_0
imath=3.1.11=hfc55251_0
importlib-metadata=7.1.0=pypi_0
jasper=4.2.4=h536e39c_0
jinja2=3.1.4=py312h06a4308_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=pypi_0
krb5=1.21.2=h659d440_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.38=h1181459_1
lerc=4.0.0=h27087fc_0
libabseil=20240116.2=cxx17_h6a678d5_0
libaec=1.1.3=h59595ed_0
libasprintf=0.22.5=h661eb56_2
libasprintf-devel=0.22.5=h661eb56_2
libass=0.17.1=h8fe9dca_1
libblas=3.9.0=22_linux64_openblas
libcap=2.69=h0f662aa_0
libcblas=3.9.0=22_linux64_openblas
libclang-cpp18.1=18.1.6=default_h127d8a8_0
libclang13=18.1.6=default_h5d6823c_0
libcups=2.3.3=h4637d8d_4
libcurl=8.8.0=hca28451_0
libdeflate=1.20=hd590300_0
libdrm=2.4.120=hd590300_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libexpat=2.6.2=h59595ed_0
libffi=3.4.4=h6a678d5_1
libflac=1.4.3=h59595ed_0
libgcc-ng=13.2.0=h77fa898_7
libgcrypt=1.10.3=hd590300_0
libgettextpo=0.22.5=h59595ed_2
libgettextpo-devel=0.22.5=h59595ed_2
libgfortran-ng=13.2.0=h69a702a_7
libgfortran5=13.2.0=hca663fb_7
libglib=2.80.2=hf974151_0
libglu=9.0.0=hf484d3e_1
libgomp=13.2.0=h77fa898_7
libgpg-error=1.49=h4f305b6_0
libhwloc=2.10.0=default_h5622ce7_1001
libiconv=1.17=hd590300_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=22_linux64_openblas
liblapacke=3.9.0=22_linux64_openblas
libllvm18=18.1.6=hb77312f_0
libmagma=2.7.2=h09b5827_2
libmagma_sparse=2.7.2=h09b5827_3
libnghttp2=1.58.0=h47da74e_1
libnsl=2.0.1=hd590300_0
libogg=1.3.5=h27cfd23_1
libopenblas=0.3.27=pthreads_h413a1c8_0
libopencv=4.9.0=qt6_py312h711d6c5_615
libopenvino=2024.1.0=h2da1b83_7
libopenvino-auto-batch-plugin=2024.1.0=hb045406_7
libopenvino-auto-plugin=2024.1.0=hb045406_7
libopenvino-hetero-plugin=2024.1.0=h5c03a75_7
libopenvino-intel-cpu-plugin=2024.1.0=h2da1b83_7
libopenvino-intel-gpu-plugin=2024.1.0=h2da1b83_7
libopenvino-intel-npu-plugin=2024.1.0=he02047a_7
libopenvino-ir-frontend=2024.1.0=h5c03a75_7
libopenvino-onnx-frontend=2024.1.0=h07e8aee_7
libopenvino-paddle-frontend=2024.1.0=h07e8aee_7
libopenvino-pytorch-frontend=2024.1.0=he02047a_7
libopenvino-tensorflow-frontend=2024.1.0=h39126c6_7
libopenvino-tensorflow-lite-frontend=2024.1.0=he02047a_7
libopus=1.3.1=h7b6447c_0
libpciaccess=0.18=hd590300_0
libpng=1.6.43=h2797004_0
libpq=16.3=ha72fbe1_0
libprotobuf=4.25.3=h08a7969_0
libsndfile=1.2.2=hc60ed4a_1
libsqlite=3.45.3=h2797004_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=13.2.0=hc0a3c3a_7
libsystemd0=255=h3516f8a_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.6.0=h1dd3fc0_3
libtorch=2.3.0=cuda118_h8db9d67_301
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libuv=1.48.0=hd590300_0
libva=2.21.0=hd590300_0
libvorbis=1.3.7=h7b6447c_0
libvpx=1.14.0=h59595ed_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.15=h7f8727e_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h662e7e4_0
libxml2=2.12.7=hc051c1a_0
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=18.1.6=ha31de31_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py312h5eee18b_0
matplotlib=3.9.0=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.2.0=h84fe81f_50496
mmcv=2.2.0=cuda118py312h93fe2d6_201
mmengine=0.10.4=pypi_0
mpg123=1.32.6=h59595ed_0
mpmath=1.3.0=py312h06a4308_0
mysql-common=8.3.0=hf1915f5_4
mysql-libs=8.3.0=hca2cd23_4
nccl=2.21.5.1=h6103f9b_0
ncurses=6.4=h6a678d5_0
nettle=3.9.1=h7ab15ed_0
networkx=3.1=py312h06a4308_0
numpy=1.26.4=py312h2809609_0
numpy-base=1.26.4=py312he1a6c75_0
ocl-icd=2.3.2=h5eee18b_1
opencv=4.9.0=qt6_py312hac6a15e_615
openexr=3.2.2=haf962dd_1
openh264=2.4.1=h59595ed_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.3.0=h4ab18f5_3
p11-kit=0.24.1=hc5aa10d_0
packaging=24.0=pypi_0
pcre2=10.43=hcad00b1_0
pillow=10.3.0=py312hdcec9eb_0
pip=24.0=py312h06a4308_0
pixman=0.43.2=h59595ed_0
platformdirs=4.2.2=pypi_0
pugixml=1.14=h59595ed_0
pulseaudio-client=17.0=hb77b528_0
py-opencv=4.9.0=qt6_py312hb24711e_615
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.12=4_cp312
pytorch=2.3.0=cuda118_py312h3690e1b_301
pyyaml=6.0.1=py312h5eee18b_0
qt6-main=6.7.1=h2471661_2
readline=8.2=h5eee18b_0
regex=2023.10.3=py312h5eee18b_0
requests=2.32.2=py312h06a4308_0
rich=13.7.1=pypi_0
safetensors=0.4.3=pypi_0
setuptools=69.5.1=py312h06a4308_0
six=1.16.0=pypi_0
sleef=3.5.1=h9b69904_2
snappy=1.2.0=hdb0a2a9_1
sqlite=3.31.1=h7b6447c_0
svt-av1=2.1.0=hac33072_0
sympy=1.12=py312h06a4308_0
tbb=2021.12.0=h297d8ca_1
termcolor=2.4.0=pypi_0
timm=1.0.3=pypi_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.1=py312h06a4308_1
torchvision=0.18.0=cuda118py312hff21f95_0
tqdm=4.66.4=pypi_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=py312h06a4308_0
wayland=1.23.0=h5291e77_0
wheel=0.43.0=py312h06a4308_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xcb-util=0.4.0=hd590300_1
xcb-util-cursor=0.1.4=hd590300_1
xcb-util-image=0.4.0=h8ee46fc_1
xcb-util-keysyms=0.4.0=h8ee46fc_1
xcb-util-renderutil=0.3.9=hd590300_1
xcb-util-wm=0.4.1=h8ee46fc_1
xkeyboard-config=2.41=hd590300_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.9=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h27cfd23_1007
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yapf=0.40.2=py312h06a4308_0
zipp=3.19.1=pypi_0
zlib=1.2.13=h4ab18f5_6
zstd=1.5.6=ha6fb4c9_0
